Use a pipe to notify Ruby thread of new ZK events.
Previous approach was to use polling with a period of 100 msec. The
new approach is to associate a pipe with the queue that we use to
send events from ZK to Ruby land. When the Ruby consumer runs out
of events to process, it does rb_thread_select() on the pipe.
Whenever the ZK side enqueues a new event, we write into the
pipe.