Avoid segfault during close
It's necessary to have a way of signaling to the dispatch thread that
it's time to shut down, and wait for that thread to exit before
invalidating the zk handle.

This patch adds a wake_event_loop! which writes a single byte to the
pipe used by the queue to signal an event. When the @_running ivar is
false and the loop is woken, it will return nil, and the dispatch
methods will know to ignore that value and return. This allows us to
shut down the dispatch thread cleanly.