create Zookeeper#selectable_io for use with EventMachine
In order to facilitate use with EventMachine, we need (at the ruby
level) to be able to wake up when an event has been delivered. This is
already happening at the C layer, so we construct a ruby IO object that
contains a dup() of the pipe's read side and make it available.