fix the bug, cannot raise an exception from outside interpreter
For some reason, a write to the queue's pipe was failing which
then called rb_raise() which totally ripped apart the fabric
of the universe. Since failing to write to the pipe is a very
rare situation and is *probably* happening during shutdown
(*probably*...we *hope*), we output a message if ZKRBDebugging is
true. Otherwise, ignore the error.

Also add a warning message about doing that again (it was bad, mm'kay).