Awesome refactor of CZookeeper, breaks ZookeeperBase
ZookeeperBase no longer has CZookeeper as a superclass. Now we do
all the lifecycle management directly on CZookeeper (some in ruby,
some in C), and create an instance of that in ZookeeperBase (like how
the java side is implemented).

This separation will allow reopen() to actually work.

Also refactored the method_get_next_event function to better handle
the shutdown case (using @_shutting_down) instead of flipping the @_running
ivar back to false, which was problematic for a number of reasons.

add a spec specifically for CZookeeper, will allow us to isolate it (and
possibly even mock it).