event delivery should stop when the client is closed.
Before the refactor, this was the behavior. This helps prevent "straggler" events
delivered after the connection has been closed. The old implementation seems to have
had slightly different event dispatch behavior (probably because it wasn't a thread
popping off a queue, rather it was directly attached to guts of the C implementation).