make assert_open more sane
Since all calls occur on the event thread (for the C module), it makes more
sense to do the check as part of the call itself inside the event thread,
rather than queue it up as a separate call.

explicit calls to 'state' will still get priority through the loop.